requirements.txt
streamlit
pdfminer.six
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
sentence-transformers
faiss-cpu
groq